API reference

Antique.CoulombTwoBody — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(\pmb{r}) = E \psi(\pmb{r}),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2\mu} \nabla^2 + \frac{z_1 z_2}{r/a_0} E_\mathrm{h},\]

where $\mu=\left(\frac{1}{m_1}+\frac{1}{m_2}\right)^{-1}$ is the reduced mass of particle 1 and particle 2. The potential includes only Coulomb interaction and it does not include fine or hyperfine interactions in this model. Parameters are specified with the following struct:

CTB = CoulombTwoBody(z₁=-1, z₂=1, m₁=1.0, m₂=1.0, mₑ=1.0, a₀=1.0, Eₕ=1.0, ℏ=1.0)

$z₁$ is the charge number of particle 1, $z₂$ is the charge number of particle 2, $m₁$ is the mass of particle 1, $m₂$ is the mass of particle 2, $m_\mathrm{e}$ is the electron mass (use the same unit as $m₁$ and $m₂$. For example of hydrogen atom, use $m_\mathrm{e}=9.1093837139\times10^{-31}\mathrm{kg}$, $m_1=9.1093837139\times10^{-31}\mathrm{kg}$ and $m_2=1.67262192595\times10^{-27}\mathrm{kg}$ in the IS unit system, use $~m_\mathrm{e}=1.0~m_\mathrm{e}$, $m_1=1.0~m_\mathrm{e}$ and $m_2=1836.152673426~m_\mathrm{e}$ in the atomic unit.), $a_0$ is the Bohr radius, $E_\mathrm{h}$ is the Hartree energy and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

The Digital Library of Mathematical Functions (DLMF), 18.3 Table1, 18.5 Table1, 18.5.16, 18.5.17

cpprefjp, assoc_legendre, assoc_laguerre

A. Messiah, Quanfum Mechanics VOLUME Ⅰ (North-Holland Publishing Company, 1961), p.412 I. THE HYDROGEN ATOM

D. J. Griffiths, D. F. Schroeter, Introduction to Quantum Mechanics Third Edition (Cambridge University Press, 2018) p.143 4.2 THE HYDROGEN ATOM, p.200 Problem 5.1, p.200 Problem 5.2

W. Greiner, Quantum Mechanics: An Introduction Forth Edition (Springer, 2001) p.217 The Hydrogen Atom, p.236 9.5 Spectrum of a Diatomic Molecule

Antique.DeltaPotential — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(x) = E \psi(x),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2m} \frac{\mathrm{d}^2}{\mathrm{d}x ^2} - \alpha \delta(x).\]

Parameters are specified with the following struct:

DP = DeltaPotential(α=1.0, m=1.0, ℏ=1.0)

$\alpha$ is the potential strength, $m$ is the mass of particle and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

D. J. Griffiths, D. F. Schroeter, Introduction to Quantum Mechanics Third Edition (Cambridge University Press, 2018) p.63, 2.5.2 The Delta-Function Well

UCSD Physics 130, Quantum Physics

Antique.HarmonicOscillator — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(x) = E \psi(x),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2m} \frac{\mathrm{d}^2}{\mathrm{d}x ^2} + \frac{1}{2} k x^2.\]

Parameters are specified with the following struct:

HO = HarmonicOscillator(k=1.0, m=1.0, ℏ=1.0)

$k$ is the force constant, $m$ is the mass of the particle and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

Main:

The Digital Library of Mathematical Functions (DLMF) 18.5.18

cpprefjp, hermite

D. J. Griffiths, D. F. Schroeter, Introduction to Quantum Mechanics Third Edition (Cambridge University Press, 2018) p.48, 2.3.2 Analytic Method

Supplemental:

The Digital Library of Mathematical Functions (DLMF) 18.3 Table1, 18.5 Table1, 18.5.13, 18.5.18

L. D. Landau, E. M. Lifshitz, Quantum Mechanics (Pergamon Press, 1965) p.595 (a.4), (a.6)

L. I. Schiff, Quantum Mechanics (McGraw-Hill Book Company, 1968) p.71 (13.12)

A. Messiah, Quanfum Mechanics (Dover Publications, 1999) p.491 (B.59)

W. Greiner, Quantum Mechanics: An Introduction Third Edition (Springer, 1994) p.152 (7.22)

D. J. Griffiths, Introduction to Quantum Mechanics (Prentice Hall, 1995) p.41 Table 2.1, p.43 (2.70)

D. A. McQuarrie, J. D. Simon, Physical Chemistry: A Molecular Approach (University Science Books, 1997) p.170 Table 5.2

P. W. Atkins, J. De Paula, Atkins' Physical Chemistry, 8th edition (W. H. Freeman, 2008) p.293 Table 9.1

J. J. Sakurai, J. Napolitano, Modern Quantum Mechanics Third Edition (Cambridge University Press, 2021) p.524 (B.29)

Antique.HydrogenAtom — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(\pmb{r}) = E \psi(\pmb{r}),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2\mu} \nabla^2 - \frac{Z}{r/a_0} E_\mathrm{h},\]

where $\mu=\left(\frac{1}{m_\mathrm{e}}+\frac{1}{m_\mathrm{p}}\right)^{-1}$ is the reduced mass of electron $\mathrm{e}$ and proton $\mathrm{p}$. $\mu = m_\mathrm{e}$ holds in the limit $m_\mathrm{p}\rightarrow\infty$. The potential includes only Coulomb interaction and it does not include fine or hyperfine interactions in this model. Parameters are specified with the following struct:

HA = HydrogenAtom(Z=1, mₑ=1.0, a₀=1.0, Eₕ=1.0, ℏ=1.0)

$Z$ is the atomic number, $m_\mathrm{e}$ is the electron mass, $a_0$ is the Bohr radius, $E_\mathrm{h}$ is the Hartree energy and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

Main:

The Digital Library of Mathematical Functions (DLMF), 18.3 Table1, 18.5 Table1, 18.5.16, 18.5.17

cpprefjp, assoc_legendre, assoc_laguerre

A. Messiah, Quanfum Mechanics VOLUME Ⅰ (North-Holland Publishing Company, 1961), p.412 (XI.3), p.419 (XI.18) (XI.18a) (XI.18b), p.483 (B.12), p.493 (B.71) (B.72), p.494 (B.81), p495 (B.93)

Supplemental:

cpprefjp, legendre, assoc_legendre, laguerre, assoc_laguerre

The Digital Library of Mathematical Functions (DLMF), 18.3 Table1, 18.5 Table1, 18.5.16, 18.5.17, 18.5.12

L. D. Landau, E. M. Lifshitz, Quantum Mechanics (Pergamon Press, 1965), p.598 (c.1), p.598 (c.4), p.603 (d.13)

L. I. Schiff, Quantum Mechanics (McGraw-Hill Book Company, 1968), p.79 (14.12), p.93 (16.19)

A. Messiah, Quanfum Mechanics (Dover Publications, 1999), p.493 (B.72), p.494 Table, p.493 (B.72), p.483 (B.12), p.483 (B.12)

W. Greiner, Quantum Mechanics: An Introduction Third Edition (Springer, 1994), p.83 (4), p.83 (5), p.149 (21)

D. J. Griffiths, Introduction to Quantum Mechanics (Prentice Hall, 1995), p.126 (4.28), p.96 Table3.1, p.126 (4.27), p.139 (4.88), p.140 Table4.4, p.139 (4.87), p.140 Table4.5

D. A. McQuarrie, J. D. Simon, Physical Chemistry: A Molecular Approach (University Science Books, 1997), p.195 Table6.1, p.196 (6.26), p.196 Table6.2, p.207 Table6.4

P. W. Atkins, J. De Paula, Atkins' Physical Chemistry, 8th edition (W. H. Freeman, 2008), p.234

J. J. Sakurai, J. Napolitano, Modern Quantum Mechanics Third Edition (Cambridge University Press, 2021), p.245 Problem 3.30.b,

Antique.InfinitePotentialWell — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(x) = E \psi(x),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2m} \frac{\mathrm{d}^2}{\mathrm{d}x ^2} + V(x).\]

Parameters are specified with the following struct:

IPW = InfinitePotentialWell(L=1.0, m=1.0, ℏ=1.0)

$L$ is the length of the box, $m$ is the mass of particle and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

D. J. Griffiths, D. F. Schroeter, Introduction to Quantum Mechanics Third Edition (Cambridge University Press, 2018) p.31, 2.2 THE INFINITE SQUARE WELL

Proofs

Eigen Functions & Eigen Values

Normalization

Antique.InfinitePotentialWell3D — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(x,y,z) = E \psi(x,y,z),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2m} \left(\frac{\partial^2}{\partial x ^2} + \frac{\partial^2}{\partial y ^2} + \frac{\partial^2}{\partial z ^2}\right) + V(x,y,z).\]

Parameters are specified with the following struct:

IPW3D = InfinitePotentialWell3D(L=[1.0,1.0,1.0], m=1.0, ℏ=1.0)

$L$ is a vector of the lengths of the box in $x$,$y$,$z$-direction, $m$ is the mass of the particle and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

D. A. McQuarrie, J. D. Simon, Physical chemistry : a molecular approach (University Science Books, 1997) p.90, 3-9. The Problem of a Particle in a Three-Dimensional Box Is a Simple Extension of the One-Dimensional Case

Antique.MorsePotential — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(r) = E \psi(r),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2\mu} \frac{\mathrm{d}^2}{\mathrm{d}r ^2} + D_\mathrm{e} \left( \mathrm{e}^{-2a(r-r_e)} - 2\mathrm{e}^{-a(r-r_e)} \right),\]

where $a = \sqrt{\frac{k}{2Dₑ}}$ is defined. Parameters are specified with the following struct:

MP = MorsePotential(rₑ=2.0, Dₑ=0.1, k=0.1, µ=918.1, ℏ=1.0)

$r_\mathrm{e}$ is the equilibrium bond distance, $D_\mathrm{e}$ is the the well depth , $k$ is the force constant, $\mu$ is the reduced mass and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

P. M. Morse, Phys. Rev., 34, 57 (1929)

J. P. Dahl, M. Springborg, J. Chem. Phys., 88, 4535 (1988). (62), (63)

W. K. Shao, Y. He, J. Pan, J. Nonlinear Sci. Appl., 9, 5, 3388 (2016). (1.6)

The Digital Library of Mathematical Functions (DLMF) 18.3 Table1, 18.5 Table1, 18.5.12, 18.5.17_5

Antique.PoschlTeller — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(x) = E \psi(x),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2 m} \frac{\mathrm{d}^2}{\mathrm{d}x ^2} - \frac{\hbar^2}{m x_0^2} \frac{\lambda(\lambda+1)}{2} \frac{1}{\mathrm{cosh}^2(x/x_0)}.\]

After introducing the dimensionless variables

\[ x^\ast \equiv x/x_0,\qquad E^\ast \equiv \frac{\hbar^2}{m x_0^2} E\]

the Schrödinger equation reduces to

\[ \hat{H}^\ast \psi(x^\ast) = E^\ast \psi(x^\ast),\]

with

\[ \hat{H}^\ast = - \frac{1}{2} \frac{\mathrm{d}^2}{\mathrm{d}{x^\ast}^2} - \frac{\lambda(\lambda+1)}{2} \frac{1}{\mathrm{cosh}^2(x^\ast)}.\]

Parameters are specified within the following struct:

PT = PoschlTeller(λ=1, m=1.0, ℏ=1.0, x₀=1.0)

$\lambda$ determines the potential strength. Currently only integer values for $\lambda$ are supported.

References

G. Pöschl, E. Teller, Zeitschrift für Physik, 83 (3–4), 143 (1933): More general definitions are given as (2a), (2b) or (11).

S. Flügge, Practical Quantum Mechanics (Springer Berlin Heidelberg, 1999) p.94 Problem 39. Potential hole of modified Poschl-Teller type.

Antique.RigidRotor — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(\theta,\varphi) = E \psi(\theta,\varphi),\]

and the Hamiltonian

\[\begin{aligned} \hat{H} &= - \frac{\hbar^2}{2\mu} \nabla^2 + V(r), \\ &= - \frac{\hbar^2}{2I} \left[ \frac{1}{\sin\theta} \frac{\partial}{\partial\theta} \left(\sin\theta \frac{\partial}{\partial\theta}\right) + \frac{1}{\sin^2\theta} \frac{\partial^2}{\partial\phi^2} \right], \\ &= \frac{L^2}{2I} \end{aligned}\]

where $I=\mu R^2$ is the moment of intertia, $\mu=\left(\frac{1}{m_1}+\frac{1}{m_2}\right)^{-1}$ is the reduced mass of two particles, $R$ is the distance between the two particles, and $L^2$ is the angular momentum operator. Parameters are specified with the following struct:

RR = RigidRotor(m₁=1.0, m₂=1.0, R=1.0, ℏ=1.0)

$m₁$ and $m₂$ are mass of two particles, $R$ is the distance, and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

D. A. McQuarrie, J. D. Simon, Physical chemistry : a molecular approach (University Science Books, 1997) p.173, 5.8 The Energy Levels of a Rigid Rotator Are $E = \hbar^2 J(J+1) / 2I$

Chemistry Libre Texts

Antique.SphericalOscillator — Type

Model

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(\pmb{r}) = E \psi(\pmb{r}),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2\mu} \nabla^2 + \frac{1}{2} k r^2.\]

Parameters are specified with the following struct:

SO = SphericalOscillator(k=1.0, μ=1.0, ℏ=1.0)

$k$ is the force constant, $μ$ is the mass of particle and $\hbar$ is the reduced Planck constant (Dirac's constant).

References

S. Flügge, Practical Quantum Mechanics (Springer Berlin Heidelberg, 1999) p.166, Problem 65. Spherical oscillator.

Quantum harmonic oscillator

UCSD Physics 130, Quantum Physics

Antique.E — Method

E(model::CoulombTwoBody; n::Int=1)

\[E_n = -\frac{(z_1 z_2)^2}{2n^2} \frac{\mu}{m_\mathrm{e}} E_\mathrm{h},\]

where $\mu=\left(\frac{1}{m_1}+\frac{1}{m_2}\right)^{-1}$ is the reduced mass of particle 1 and particle 2, $E_\mathrm{h} = \frac{\hbar^2}{m_\mathrm{e}{a_0}^2} = \frac{e^2}{4\pi\varepsilon_0a_0} = \frac{m_\mathrm{e}e^4}{\left(4\pi\varepsilon_0\right)^2\hbar^2}$ is the Hartree energy, one of atomic unit. About atomic units, see section 3.9.2 of the IUPAC GreenBook. In other units, $E_\mathrm{h} = 27.211~386~245~988(53)~\mathrm{eV}$ from here.

Antique.E — Method

E(model::DeltaPotential)

\[E = - \frac{m\alpha^2}{2\hbar^2}\]

Antique.E — Method

E(model::HarmonicOscillator; n::Int=0)

\[E_n = \hbar \omega \left( n + \frac{1}{2} \right),\]

where $\omega = \sqrt{k/m}$ is the angular frequency.

Antique.E — Method

E(model::HydrogenAtom; n::Int=1)

\[E_n = -\frac{m_\mathrm{e} e^4 Z^2}{2n^2(4\pi\varepsilon_0)^2\hbar^2} = -\frac{Z^2}{2n^2} E_\mathrm{h},\]

where $E_\mathrm{h} = \frac{\hbar^2}{m_\mathrm{e}{a_0}^2} = \frac{e^2}{4\pi\varepsilon_0a_0} = \frac{m_\mathrm{e}e^4}{\left(4\pi\varepsilon_0\right)^2\hbar^2}$ is the Hartree energy, one of atomic unit. About atomic units, see section 3.9.2 of the IUPAC GreenBook. In other units, $E_\mathrm{h} = 27.211~386~245~988(53)~\mathrm{eV}$ from here.

Antique.E — Method

E(model::InfinitePotentialWell3D; n::Vector{Int}=[1,1,1])

\[E_{n_x,n_y,n_z} = \frac{\hbar^2 n_x^2 \pi^2}{2 m L_x^2} + \frac{\hbar^2 n_y^2 \pi^2}{2 m L_y^2} + \frac{\hbar^2 n_z^2 \pi^2}{2 m L_z^2}\]

Antique.E — Method

E(model::InfinitePotentialWell; n::Int=1)

\[E_n = \frac{\hbar^2 n^2 \pi^2}{2 m L^2}\]

Antique.E — Method

E(model::MorsePotential; n::Int=0, nocheck=false)

\[E_n = - D_\mathrm{e} + \hbar \omega \left( n + \frac{1}{2} \right) - \chi \hbar \omega \left( n + \frac{1}{2} \right)^2,\]

where $\omega = \sqrt{k/µ}$ and $\chi = \frac{\hbar\omega}{4D_\mathrm{e}}$ are defined.

Antique.E — Method

E(model::PoschlTeller; n::Int=0, nocheck=false)

\[E_n = -\frac{\hbar^2}{m x_0^2}\frac{\mu^2}{2},\]

where $\mu = \mu(n) = n_\mathrm{max}-n+1$, and $n_\mathrm{max} = \left\lfloor \lambda \right\rfloor - 1$.

Antique.E — Method

E(model::RigidRotor; l::Int=0)

\[E_l = \frac{\hbar^2}{2I}l(l+1),\]

where $I=\mu R^2$ is the moment of inertia, $R$ is the distance, and $\mu$ is the reduced mass of the two particles.

Antique.E — Method

E(model::SphericalOscillator; n::Int=0, l::Int=0)

\[E_{nl} = \left(2n + l + \frac{3}{2}\right)\hbar \omega,\]

where $\omega = \sqrt{k/\mu}$.

Antique.H — Method

H(model::HarmonicOscillator, x; n=0)

Rodrigues' formula & closed-form:

\[\begin{aligned} H_{n}(x) &:= (-1)^n \mathrm{e}^{x^2} \frac{\mathrm{d}^n}{\mathrm{d}x^n} \mathrm{e}^{-x^2} \\ &= n! \sum_{m=0}^{\lfloor n/2 \rfloor} \frac{(-1)^m}{m! (n-2m)!}(2 x)^{n-2m}. \end{aligned}\]

Examples:

\[\begin{aligned} H_{0}(x) &= 1, \\ H_{1}(x) &= 2 x, \\ H_{2}(x) &= -2 + 4 x^{2}, \\ H_{3}(x) &= -12 x + 8 x^{3}, \\ H_{4}(x) &= 12 - 48 x^{2} + 16 x^{4}, \\ H_{5}(x) &= 120 x - 160 x^{3} + 32 x^{5}, \\ H_{6}(x) &= -120 + 720 x^{2} - 480 x^{4} + 64 x^{6}, \\ H_{7}(x) &= -1680 x + 3360 x^{3} - 1344 x^{5} + 128 x^{7}, \\ H_{8}(x) &= 1680 - 13440 x^{2} + 13440 x^{4} - 3584 x^{6} + 256 x^{8}, \\ H_{9}(x) &= 30240 x - 80640 x^{3} + 48384 x^{5} - 9216 x^{7} + 512 x^{9}, \\ &\vdots \end{aligned}\]

Antique.L — Method

L(model::CoulombTwoBody, x; n=0, k=0)

Note

The associated Laguerre polynomials $L_n^{k}(x)$, not the generalized Laguerre polynomials $L_n^{(\alpha)}(x)$, are used in this model.

Rodrigues' formula & closed-form:

\[\begin{aligned} L_n^{k}(x) &= \frac{\mathrm{d}^k}{\mathrm{d}x^k} L_n(x) \\ &= \frac{\mathrm{d}^k}{\mathrm{d}x^k} \frac{1}{n!} \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right) \\ &= \sum_{m=0}^{n-k} (-1)^{m+k} \frac{n!}{m!(m+k)!(n-m-k)!} x^m \\ &= (-1)^k L_{n-k}^{(k)}(x), \end{aligned}\]

where Laguerre polynomials are defined as $L_n(x)=\frac{1}{n!}\mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$.

Examples:

\[\begin{aligned} L_0^0(x) &= 1, \\ L_1^0(x) &= 1 - x, \\ L_1^1(x) &= 1, \\ L_2^0(x) &= 1 - 2 x + 1/2 x^2, \\ L_2^1(x) &= 2 - x, \\ L_2^2(x) &= 1, \\ L_3^0(x) &= 1 - 3 x + 3/2 x^2 - 1/6 x^3, \\ L_3^1(x) &= 3 - 3 x + 1/2 x^2, \\ L_3^2(x) &= 3 - x, \\ L_3^3(x) &= 1, \\ L_4^0(x) &= 1 - 4 x + 3 x^2 - 2/3 x^3 + 5/12 x^4, \\ L_4^1(x) &= 4 - 6 x + 2 x^2 - 1/6 x^3, \\ L_4^2(x) &= 6 - 4 x + 1/2 x^2, \\ L_4^3(x) &= 4 - x, \\ L_4^4(x) &= 1, \\ \vdots \end{aligned}\]

Antique.L — Method

L(model::HydrogenAtom, x; n=0, k=0)

Note

The associated Laguerre polynomials $L_n^{k}(x)$, not the generalized Laguerre polynomials $L_n^{(\alpha)}(x)$, are used in this model.

Rodrigues' formula & closed-form:

\[\begin{aligned} L_n^{k}(x) &= \frac{\mathrm{d}^k}{\mathrm{d}x^k} L_n(x) \\ &= \frac{\mathrm{d}^k}{\mathrm{d}x^k} \frac{1}{n!} \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right) \\ &= \sum_{m=0}^{n-k} (-1)^{m+k} \frac{n!}{m!(m+k)!(n-m-k)!} x^m \\ &= (-1)^k L_{n-k}^{(k)}(x), \end{aligned}\]

where Laguerre polynomials are defined as $L_n(x)=\frac{1}{n!}\mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$.

Examples:

\[\begin{aligned} L_0^0(x) &= 1, \\ L_1^0(x) &= 1 - x, \\ L_1^1(x) &= 1, \\ L_2^0(x) &= 1 - 2 x + 1/2 x^2, \\ L_2^1(x) &= 2 - x, \\ L_2^2(x) &= 1, \\ L_3^0(x) &= 1 - 3 x + 3/2 x^2 - 1/6 x^3, \\ L_3^1(x) &= 3 - 3 x + 1/2 x^2, \\ L_3^2(x) &= 3 - x, \\ L_3^3(x) &= 1, \\ L_4^0(x) &= 1 - 4 x + 3 x^2 - 2/3 x^3 + 5/12 x^4, \\ L_4^1(x) &= 4 - 6 x + 2 x^2 - 1/6 x^3, \\ L_4^2(x) &= 6 - 4 x + 1/2 x^2, \\ L_4^3(x) &= 4 - x, \\ L_4^4(x) &= 1, \\ \vdots \end{aligned}\]

Antique.L — Method

L(model::MorsePotential, x; n=0, α=0)

Note

The generalized Laguerre polynomials $L_n^{(\alpha)}(x)$, not the associated Laguerre polynomials $L_n^{k}(x)$, are used in this model.

Rodrigues' formula & closed-form:

\[\begin{aligned} L_n^{(\alpha)}(x) &= \frac{x^{-\alpha}e^x}{n!} \frac{d^n}{dx^n}\left(x^{n+\alpha}e^{-x}\right) \\ &= \sum_{k=0}^n(-1)^k \left(\begin{array}{l} n+\alpha \\ n-k \end{array}\right) \frac{x^k}{k !} \\ &= \sum_{k=0}^n(-1)^k \frac{\Gamma(\alpha+n+1)}{\Gamma(\alpha+k+1)\Gamma(n-k+1)} \frac{x^k}{k !}. \end{aligned}\]

Examples:

\[\begin{aligned} L_0^{(0)}(x) &= 1, \\ L_1^{(0)}(x) &= 1 - x, \\ L_1^{(1)}(x) &= 2 - x, \\ L_2^{(0)}(x) &= 1 - 2 x + 1/2 x^{2}, \\ L_2^{(1)}(x) &= 3 - 3 x + 1/2 x^{2}, \\ L_2^{(2)}(x) &= 6 - 4 x + 1/2 x^{2}, \\ L_3^{(0)}(x) &= 1 - 3 x + 3/2 x^{2} - 1/6 x^{3}, \\ L_3^{(1)}(x) &= 4 - 6 x + 2 x^{2} - 1/6 x^{3}, \\ L_3^{(2)}(x) &= 10 - 10 x + 5/2 x^{2} - 1/6 x^{3}, \\ L_3^{(3)}(x) &= 20 - 15 x + 3 x^{2} - 1/6 x^{3}, \\ L_4^{(0)}(x) &= 1 - 4 x + 3 x^{2} - 2/3 x^{3} + 1/24 x^{4}, \\ L_4^{(1)}(x) &= 5 - 10 x + 5 x^{2} - 5/6 x^{3} + 1/24 x^{4}, \\ L_4^{(2)}(x) &= 15 - 20 x + 15/2 x^{2} - 1 x^{3} + 1/24 x^{4}, \\ L_4^{(3)}(x) &= 35 - 35 x + 21/2 x^{2} - 7/6 x^{3} + 1/24 x^{4}, \\ L_4^{(4)}(x) &= 70 - 56 x + 14 x^{2} - 4/3 x^{3} + 1/24 x^{4}, \\ \vdots \end{aligned}\]

Antique.L — Method

L(model::SphericalOscillator, x; n=0, α=0)

Note

The generalized Laguerre polynomials $L_n^{(\alpha)}(x)$, not the associated Laguerre polynomials $L_n^{k}(x)$, are used in this model.

Rodrigues' formula & closed-form:

\[\begin{aligned} L_n^{(\alpha)}(x) &= \frac{x^{-\alpha}e^x}{n!} \frac{d^n}{dx^n}\left(x^{n+\alpha}e^{-x}\right) \\ &= \sum_{k=0}^n(-1)^k \left(\begin{array}{l} n+\alpha \\ n-k \end{array}\right) \frac{x^k}{k !} \\ &= \sum_{k=0}^n(-1)^k \frac{\Gamma(\alpha+n+1)}{\Gamma(\alpha+k+1)\Gamma(n-k+1)} \frac{x^k}{k !}. \end{aligned}\]

Examples:

\[\begin{aligned} L_0^{(0)}(x) &= 1, \\ L_1^{(0)}(x) &= 1 - x, \\ L_1^{(1)}(x) &= 2 - x, \\ L_2^{(0)}(x) &= 1 - 2 x + 1/2 x^{2}, \\ L_2^{(1)}(x) &= 3 - 3 x + 1/2 x^{2}, \\ L_2^{(2)}(x) &= 6 - 4 x + 1/2 x^{2}, \\ L_3^{(0)}(x) &= 1 - 3 x + 3/2 x^{2} - 1/6 x^{3}, \\ L_3^{(1)}(x) &= 4 - 6 x + 2 x^{2} - 1/6 x^{3}, \\ L_3^{(2)}(x) &= 10 - 10 x + 5/2 x^{2} - 1/6 x^{3}, \\ L_3^{(3)}(x) &= 20 - 15 x + 3 x^{2} - 1/6 x^{3}, \\ L_4^{(0)}(x) &= 1 - 4 x + 3 x^{2} - 2/3 x^{3} + 1/24 x^{4}, \\ L_4^{(1)}(x) &= 5 - 10 x + 5 x^{2} - 5/6 x^{3} + 1/24 x^{4}, \\ L_4^{(2)}(x) &= 15 - 20 x + 15/2 x^{2} - 1 x^{3} + 1/24 x^{4}, \\ L_4^{(3)}(x) &= 35 - 35 x + 21/2 x^{2} - 7/6 x^{3} + 1/24 x^{4}, \\ L_4^{(4)}(x) &= 70 - 56 x + 14 x^{2} - 4/3 x^{3} + 1/24 x^{4}, \\ \vdots \end{aligned}\]

Antique.P — Method

P(model::CoulombTwoBody, x; n=0, m=0)

Rodrigues' formula & closed-form:

\[\begin{aligned} P_n^m(x) &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} P_n(x) \\ &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right] \\ &= \frac{1}{2^n} (1-x^2)^{m/2} \sum_{j=0}^{\left\lfloor\frac{n-m}{2}\right\rfloor} (-1)^j \frac{(2n-2j)!}{j! (n-j)! (n-2j-m)!} x^{(n-2j-m)}. \end{aligned},\]

where Legendre polynomials are defined as $P_n(x) = \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right]$. Note that $P_l^{-m} = (-1)^m \frac{(l-m)!}{(l+m)!} P_l^m$ for $m<0$. (It is not compatible with $P_k^m(t) = (-1)^m\left( 1-t^2 \right)^{m/2} \frac{\mathrm{d}^m P_k(t)}{\mathrm{d}t^m}$ caused by $(-1)^m$.) The specific formulae are given below.

Examples:

\[\begin{aligned} P_{0}^{0}(x) &= 1, \\ P_{1}^{0}(x) &= x, \\ P_{1}^{1}(x) &= \left(+1\right)\sqrt{1-x^2}, \\ P_{2}^{0}(x) &= -1/2 + 3/2 x^{2}, \\ P_{2}^{1}(x) &= \left(-3 x\right)\sqrt{1-x^2}, \\ P_{2}^{2}(x) &= 3 - 6 x, \\ P_{3}^{0}(x) &= -3/2 x + 5/2 x^{3}, \\ P_{3}^{1}(x) &= \left(3/2 - 15/2 x^{2}\right)\sqrt{1-x^2}, \\ P_{3}^{2}(x) &= 15 x - 30 x^{2}, \\ P_{3}^{3}(x) &= \left(15 - 30 x\right)\sqrt{1-x^2}, \\ P_{4}^{0}(x) &= 3/8 - 15/4 x^{2} + 35/8 x^{4}, \\ P_{4}^{1}(x) &= \left(- 15/2 x + 35/2 x^{3}\right)\sqrt{1-x^2}, \\ P_{4}^{2}(x) &= -15/2 + 15 x + 105/2 x^{2} - 105 x^{3}, \\ P_{4}^{3}(x) &= \left(105 x - 210 x^{2}\right)\sqrt{1-x^2}, \\ P_{4}^{4}(x) &= 105 - 420 x + 420 x^{2}, \\ & \vdots \end{aligned}\]

Antique.P — Method

P(model::HydrogenAtom, x; n=0, m=0)

Rodrigues' formula & closed-form:

\[\begin{aligned} P_n^m(x) &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} P_n(x) \\ &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right] \\ &= \frac{1}{2^n} (1-x^2)^{m/2} \sum_{j=0}^{\left\lfloor\frac{n-m}{2}\right\rfloor} (-1)^j \frac{(2n-2j)!}{j! (n-j)! (n-2j-m)!} x^{(n-2j-m)}. \end{aligned},\]

where Legendre polynomials are defined as $P_n(x) = \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right]$. Note that $P_l^{-m} = (-1)^m \frac{(l-m)!}{(l+m)!} P_l^m$ for $m<0$. (It is not compatible with $P_k^m(t) = (-1)^m\left( 1-t^2 \right)^{m/2} \frac{\mathrm{d}^m P_k(t)}{\mathrm{d}t^m}$ caused by $(-1)^m$.) The specific formulae are given below.

Examples:

\[\begin{aligned} P_{0}^{0}(x) &= 1, \\ P_{1}^{0}(x) &= x, \\ P_{1}^{1}(x) &= \left(+1\right)\sqrt{1-x^2}, \\ P_{2}^{0}(x) &= -1/2 + 3/2 x^{2}, \\ P_{2}^{1}(x) &= \left(-3 x\right)\sqrt{1-x^2}, \\ P_{2}^{2}(x) &= 3 - 6 x, \\ P_{3}^{0}(x) &= -3/2 x + 5/2 x^{3}, \\ P_{3}^{1}(x) &= \left(3/2 - 15/2 x^{2}\right)\sqrt{1-x^2}, \\ P_{3}^{2}(x) &= 15 x - 30 x^{2}, \\ P_{3}^{3}(x) &= \left(15 - 30 x\right)\sqrt{1-x^2}, \\ P_{4}^{0}(x) &= 3/8 - 15/4 x^{2} + 35/8 x^{4}, \\ P_{4}^{1}(x) &= \left(- 15/2 x + 35/2 x^{3}\right)\sqrt{1-x^2}, \\ P_{4}^{2}(x) &= -15/2 + 15 x + 105/2 x^{2} - 105 x^{3}, \\ P_{4}^{3}(x) &= \left(105 x - 210 x^{2}\right)\sqrt{1-x^2}, \\ P_{4}^{4}(x) &= 105 - 420 x + 420 x^{2}, \\ & \vdots \end{aligned}\]

Antique.P — Method

P(model::PoschlTeller, x; n=0, m=0)

Associated Legendre polynomials are the associated Legendre functions for integer indices. Here we use the same notation of the associated Legendre functions as in the model HydrogenAtom.

\[\begin{aligned} P_n^m(x) &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} P_n(x) \\ &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right] \\ &= \frac{1}{2^n} (1-x^2)^{m/2} \sum_{j=0}^{\left\lfloor\frac{n-m}{2}\right\rfloor} (-1)^j \frac{(2n-2j)!}{j! (n-j)! (n-2j-m)!} x^{(n-2j-m)}. \end{aligned}\]

Antique.P — Method

P(model::RigidRotor, x; n=0, m=0)

Rodrigues' formula & closed-form:

\[\begin{aligned} P_n^m(x) &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} P_n(x) \\ &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right] \\ &= \frac{1}{2^n} (1-x^2)^{m/2} \sum_{j=0}^{\left\lfloor\frac{n-m}{2}\right\rfloor} (-1)^j \frac{(2n-2j)!}{j! (n-j)! (n-2j-m)!} x^{(n-2j-m)}. \end{aligned},\]

where Legendre polynomials are defined as $P_n(x) = \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right]$. Note that $P_l^{-m} = (-1)^m \frac{(l-m)!}{(l+m)!} P_l^m$ for $m<0$. (It is not compatible with $P_k^m(t) = (-1)^m\left( 1-t^2 \right)^{m/2} \frac{\mathrm{d}^m P_k(t)}{\mathrm{d}t^m}$ caused by $(-1)^m$.) The specific formulae are given below.

Examples:

\[\begin{aligned} P_{0}^{0}(x) &= 1, \\ P_{1}^{0}(x) &= x, \\ P_{1}^{1}(x) &= \left(+1\right)\sqrt{1-x^2}, \\ P_{2}^{0}(x) &= -1/2 + 3/2 x^{2}, \\ P_{2}^{1}(x) &= \left(-3 x\right)\sqrt{1-x^2}, \\ P_{2}^{2}(x) &= 3 - 6 x, \\ P_{3}^{0}(x) &= -3/2 x + 5/2 x^{3}, \\ P_{3}^{1}(x) &= \left(3/2 - 15/2 x^{2}\right)\sqrt{1-x^2}, \\ P_{3}^{2}(x) &= 15 x - 30 x^{2}, \\ P_{3}^{3}(x) &= \left(15 - 30 x\right)\sqrt{1-x^2}, \\ P_{4}^{0}(x) &= 3/8 - 15/4 x^{2} + 35/8 x^{4}, \\ P_{4}^{1}(x) &= \left(- 15/2 x + 35/2 x^{3}\right)\sqrt{1-x^2}, \\ P_{4}^{2}(x) &= -15/2 + 15 x + 105/2 x^{2} - 105 x^{3}, \\ P_{4}^{3}(x) &= \left(105 x - 210 x^{2}\right)\sqrt{1-x^2}, \\ P_{4}^{4}(x) &= 105 - 420 x + 420 x^{2}, \\ & \vdots \end{aligned}\]

Antique.P — Method

P(model::SphericalOscillator, x; n=0, m=0)

Rodrigues' formula & closed-form:

\[\begin{aligned} P_n^m(x) &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} P_n(x) \\ &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right] \\ &= \frac{1}{2^n} (1-x^2)^{m/2} \sum_{j=0}^{\left\lfloor\frac{n-m}{2}\right\rfloor} (-1)^j \frac{(2n-2j)!}{j! (n-j)! (n-2j-m)!} x^{(n-2j-m)}. \end{aligned},\]

where Legendre polynomials are defined as $P_n(x) = \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right]$. Note that $P_l^{-m} = (-1)^m \frac{(l-m)!}{(l+m)!} P_l^m$ for $m<0$. (It is not compatible with $P_k^m(t) = (-1)^m\left( 1-t^2 \right)^{m/2} \frac{\mathrm{d}^m P_k(t)}{\mathrm{d}t^m}$ caused by $(-1)^m$.) The specific formulae are given below.

Examples:

\[\begin{aligned} P_{0}^{0}(x) &= 1, \\ P_{1}^{0}(x) &= x, \\ P_{1}^{1}(x) &= \left(+1\right)\sqrt{1-x^2}, \\ P_{2}^{0}(x) &= -1/2 + 3/2 x^{2}, \\ P_{2}^{1}(x) &= \left(-3 x\right)\sqrt{1-x^2}, \\ P_{2}^{2}(x) &= 3 - 6 x, \\ P_{3}^{0}(x) &= -3/2 x + 5/2 x^{3}, \\ P_{3}^{1}(x) &= \left(3/2 - 15/2 x^{2}\right)\sqrt{1-x^2}, \\ P_{3}^{2}(x) &= 15 x - 30 x^{2}, \\ P_{3}^{3}(x) &= \left(15 - 30 x\right)\sqrt{1-x^2}, \\ P_{4}^{0}(x) &= 3/8 - 15/4 x^{2} + 35/8 x^{4}, \\ P_{4}^{1}(x) &= \left(- 15/2 x + 35/2 x^{3}\right)\sqrt{1-x^2}, \\ P_{4}^{2}(x) &= -15/2 + 15 x + 105/2 x^{2} - 105 x^{3}, \\ P_{4}^{3}(x) &= \left(105 x - 210 x^{2}\right)\sqrt{1-x^2}, \\ P_{4}^{4}(x) &= 105 - 420 x + 420 x^{2}, \\ & \vdots \end{aligned}\]

Antique.R — Method

R(model::CoulombTwoBody, r; n=1, l=0)

\[R_{nl}(r) = -\sqrt{\frac{(n-l-1)!}{2n(n+l)!} \left(\frac{2Z}{n a_\mu}\right)^3} \left(\frac{2Zr}{n a_\mu}\right)^l \exp \left(-\frac{Zr}{n a_\mu}\right) L_{n+l}^{2l+1} \left(\frac{2Zr}{n a_\mu}\right),\]

where $\frac{1}{\mu} = \frac{1}{m_1}+\frac{1}{m_2}$, $a_\mu = a_0 \frac{m_\mathrm{e}}{\mu}$, $Z = - z_1 z_2$, the Laguerre polynomials are defined as $L_n(x) = \frac{1}{n!} \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$, and the associated Laguerre polynomials are defined as $L_n^{k}(x) = \frac{\mathrm{d}^k}{\mathrm{d}x^k} L_n(x)$. Note that replace $2n(n+l)!$ with $2n[(n+l)!]^3$ if the Laguerre polynomials are defined as $L_n(x) = \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$. Note that replace $L_{n+l}^{2l+1}(x)$ with $- L_{n-l-1}^{2l+1}(x)$ if the associated Laguerre polynomials are defined as $L_n^{k}(x) = (-1)^k \frac{\mathrm{d}^k}{\mathrm{d}x^k} L_{n+k}(x)$, which we call the generalized Laguerre polynomials. The domain is $0\leq r \lt \infty$.

Antique.R — Method

R(model::HydrogenAtom, r; n=1, l=0)

\[R_{nl}(r) = -\sqrt{\frac{(n-l-1)!}{2n(n+l)!} \left(\frac{2Z}{n a_0}\right)^3} \left(\frac{2Zr}{n a_0}\right)^l \exp \left(-\frac{Zr}{n a_0}\right) L_{n+l}^{2l+1} \left(\frac{2Zr}{n a_0}\right),\]

where the Laguerre polynomials are defined as $L_n(x) = \frac{1}{n!} \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$, and the associated Laguerre polynomials are defined as $L_n^{k}(x) = \frac{\mathrm{d}^k}{\mathrm{d}x^k} L_n(x)$. Note that replace $2n(n+l)!$ with $2n[(n+l)!]^3$ if the Laguerre polynomials are defined as $L_n(x) = \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$. Note that replace $L_{n+l}^{2l+1}(x)$ with $- L_{n-l-1}^{2l+1}(x)$ if the associated Laguerre polynomials are defined as $L_n^{k}(x) = (-1)^k \frac{\mathrm{d}^k}{\mathrm{d}x^k} L_{n+k}(x)$, which we call the generalized Laguerre polynomials. The domain is $0\leq r \lt \infty$.

Antique.R — Method

R(model::SphericalOscillator, r; n=0, l=0)

\[R_{nl}(r) = \sqrt{ \frac{\gamma^{3/2}}{2\sqrt{\pi}}} \sqrt{\frac{2^{n+l+3} n!}{(2n+2l+1)!!}} \xi^l \exp\left(-\xi^2/2\right)L_{n}^{(l+\frac{1}{2})} \left(\xi^2\right),\]

where $\gamma = \mu\omega/\hbar$ and $\xi = \sqrt{\gamma}r = \sqrt{\mu\omega/\hbar}r$ are defined. The generalized Laguerre polynomials are defined as $L_n^{(\alpha)}(x) = \frac{x^{-\alpha} \mathrm{e}^x}{n !} \frac{\mathrm{d}^n}{\mathrm{d} x^n}\left(\mathrm{e}^{-x} x^{n+\alpha}\right)$. The domain is $0\leq r \lt \infty$.

Antique.V — Method

V(model::CoulombTwoBody, r)

\[\begin{aligned} V(r) &= - \frac{z_1 z_2 e^2}{4\pi\varepsilon_0 r} &= - \frac{e^2}{4\pi\varepsilon_0 a_0} \frac{z_1 z_2}{r/a_0} &= - \frac{z_1 z_2}{r/a_0} E_\mathrm{h}, \end{aligned}\]

where $E_\mathrm{h} = \frac{\hbar^2}{m_\mathrm{e}{a_0}^2} = \frac{e^2}{4\pi\varepsilon_0a_0} = \frac{m_\mathrm{e}e^4}{\left(4\pi\varepsilon_0\right)^2\hbar^2}$ is the Hartree energy, one of atomic unit. The domain is $0\leq r \lt \infty$.

Antique.V — Method

V(model::DeltaPotential, x)

\[V(x) = -\alpha \delta(x).\]

Antique.V — Method

V(model::HarmonicOscillator, x)

\[V(x) = \frac{1}{2} k x^2 = \frac{1}{2} m \omega^2 x^2 = \frac{1}{2} \hbar \omega \xi^2,\]

where $\omega = \sqrt{k/m}$ is the angular frequency and $\xi = \sqrt{\frac{m\omega}{\hbar}}x$.

Antique.V — Method

V(model::HydrogenAtom, r)

\[\begin{aligned} V(r) &= - \frac{Ze^2}{4\pi\varepsilon_0 r} &= - \frac{e^2}{4\pi\varepsilon_0 a_0} \frac{Z}{r/a_0} &= - \frac{Z}{r/a_0} E_\mathrm{h}, \end{aligned}\]

where $E_\mathrm{h} = \frac{\hbar^2}{m_\mathrm{e}{a_0}^2} = \frac{e^2}{4\pi\varepsilon_0a_0} = \frac{m_\mathrm{e}e^4}{\left(4\pi\varepsilon_0\right)^2\hbar^2}$ is the Hartree energy, one of atomic unit. The domain is $0\leq r \lt \infty$.

Antique.V — Method

V(model::InfinitePotentialWell; x)

\[V(x) = \left\{ \begin{array}{ll} 0 & 0 \leq x \leq L \\ \infty & x \lt 0, L \lt x \end{array} \right.\]

Antique.V — Method

V(model::InfinitePotentialWell3D, x)

\[V(x,y,z) = \left\{ \begin{array}{ll} 0 & 0 \leq x \leq L_x \ \mathrm{and}\ 0 \leq y \leq L_y \ \mathrm{and}\ 0 \leq z \leq L_z \\ \infty & \mathrm{elsewhere} \end{array} \right.\]

Antique.V — Method

V(model::MorsePotential, r)

\[V(r) = D_\mathrm{e} \left( \mathrm{e}^{-2a(r-r_e)} - 2\mathrm{e}^{-a(r-r_e)} \right),\]

where $a = \sqrt{\frac{k}{2Dₑ}}$ is defined. The domain is $0\leq r \lt \infty$.

Antique.V — Method

V(model::PoschlTeller, x)

\[\begin{aligned} V(x) &= -\frac{\hbar^2}{m x_0^2} \frac{\lambda(\lambda+1)}{2} \mathrm{sech}^2(x/x_0) \\ &= -\frac{\hbar^2}{m x_0^2} \frac{\lambda(\lambda+1)}{2} \frac{1}{\mathrm{cosh}^2(x/x_0)} . \end{aligned}\]

Antique.V — Method

V(model::RigidRotor, r)

\[\begin{aligned} V(r) &= 0, \end{aligned}\]

Antique.V — Method

V(model::SphericalOscillator, r)

\[V(r) = \frac{1}{2} k r^2 = \frac{1}{2} \mu \omega^2 r^2 = \frac{1}{2} \hbar \omega \xi^2,\]

where $\omega = \sqrt{k/\mu}$ is the angular frequency and $\xi = \sqrt{\frac{\mu\omega}{\hbar}}r$.

Antique.Y — Method

Y(model::CoulombTwoBody, θ, φ; l=0, m=0)

\[Y_{lm}(\theta,\varphi) = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{2l+1}{4\pi} \frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} (\cos\theta) \mathrm{e}^{im\varphi}.\]

The domain is $0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$. Note that some variants are connected by

\[i^{|m|+m} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^m \sqrt{\frac{(l-m)!}{(l+m)!}} P_l^{m}.\]

Antique.Y — Method

Y(model::HydrogenAtom, θ, φ; l=0, m=0)

\[Y_{lm}(\theta,\varphi) = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{2l+1}{4\pi} \frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} (\cos\theta) \mathrm{e}^{im\varphi}.\]

The domain is $0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$. Note that some variants are connected by

\[i^{|m|+m} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^m \sqrt{\frac{(l-m)!}{(l+m)!}} P_l^{m}.\]

Antique.Y — Method

Y(model::RigidRotor, θ, φ; l=0, m=0)

\[Y_{lm}(\theta,\varphi) = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{2l+1}{4\pi} \frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} (\cos\theta) \mathrm{e}^{im\varphi}.\]

The domain is $0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$. Note that some variants are connected by

\[i^{|m|+m} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^m \sqrt{\frac{(l-m)!}{(l+m)!}} P_l^{m}.\]

Antique.Y — Method

Y(model::SphericalOscillator, θ, φ; l=0, m=0)

\[Y_{lm}(\theta,\varphi) = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{2l+1}{4\pi} \frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} (\cos\theta) \mathrm{e}^{im\varphi}.\]

The domain is $0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$. Note that some variants are connected by

\[i^{|m|+m} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^m \sqrt{\frac{(l-m)!}{(l+m)!}} P_l^{m}.\]

Antique.nₘₐₓ — Method

nₘₐₓ(model::MorsePotential)

Note

Note that the number of bound states is equal to the maximum quantum number nₘₐₓ, since we count the ground state from n=1 in this model.

\[n_\mathrm{max} = \left\lfloor \frac{2 D_e - \omega}{\omega} \right\rfloor,\]

where $\omega = \sqrt{k/µ}$ is defined.

Antique.nₘₐₓ — Method

nₘₐₓ(model::PoschlTeller)

Note

Note that the number of bound states nₘₐₓ + 1 is not equal to the maximum quantum number nₘₐₓ, since we count the ground state from n=0 in this model.

\[n_\mathrm{max} = \left\lfloor \lambda \right\rfloor - 1.\]

Antique.ψ — Method

ψ(model::CoulombTwoBody, r, θ, φ; n::Int=1, l::Int=0, m::Int=0)

\[\psi_{nlm}(\pmb{r}) = R_{nl}(r) Y_{lm}(\theta,\varphi)\]

The domain is $0\leq r \lt \infty, 0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$.

Antique.ψ — Method

ψ(model::DeltaPotential, x)

\[\psi(x) = \frac{\sqrt{m\alpha}}{\hbar} \mathrm{e}^{-m\alpha |x|/\hbar^2}\]

Antique.ψ — Method

ψ(model::HarmonicOscillator, x; n::Int=0)

\[\psi_n(x) = A_n H_n(\xi) \exp{\left( -\frac{\xi^2}{2} \right)},\]

where $\omega = \sqrt{k/m}$, $\xi = \sqrt{\frac{m\omega}{\hbar}}x$, $A_n = \sqrt{\frac{1}{n! 2^n} \sqrt{\frac{m\omega}{\pi\hbar}}}$, $H_n(x) = (-1)^n \mathrm{e}^{x^2} \frac{\mathrm{d}^n}{\mathrm{d}x^n} \mathrm{e}^{-x^2}$ are defined.

Antique.ψ — Method

ψ(model::HydrogenAtom, r, θ, φ; n::Int=1, l::Int=0, m::Int=0)

\[\psi_{nlm}(\pmb{r}) = R_{nl}(r) Y_{lm}(\theta,\varphi)\]

The domain is $0\leq r \lt \infty, 0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$.

Antique.ψ — Method

ψ(model::InfinitePotentialWell, x; n::Int=1)

\[\psi_n(x) = \sqrt{\frac{2}{L}} \sin \frac{n\pi x}{L}\]

Antique.ψ — Method

ψ(model::InfinitePotentialWell3D, x; n::Vector{Int}=[1,1,1])

The wave functions can be expressed as products of wave functions in a one-dimensional box.

\[\begin{aligned} \psi_{n_x,n_y,n_z}(x,y,z) &= \psi_{n_x}(x) \times \psi_{n_y}(y) \times \psi_{n_z}(z) \\ &= \sqrt{\frac{2}{L_x}} \sin \frac{n_x \pi x}{L_x} \times \sqrt{\frac{2}{L_y}} \sin \frac{n_y \pi y}{L_y} \times \sqrt{\frac{2}{L_z}} \sin \frac{n_z \pi z}{L_z} \end{aligned}\]

Antique.ψ — Method

ψ(model::MorsePotential, r; n::Int=0)

\[\psi_n(r) = N_n z^{\lambda-n-1/2} \mathrm{e}^{-z/2} L_n^{(2\lambda-2n-1)}(\xi),\]

$N_n = \sqrt{\frac{n!(2\lambda-2n-1)a}{\Gamma(2\lambda-n)}}$, $\lambda = \frac{\sqrt{2\mu D_\mathrm{e}}}{a\hbar}$, $a = \sqrt{\frac{k}{2Dₑ}}$, $L_n^{(\alpha)}(x) = \frac{x^{-\alpha} \mathrm{e}^x}{n !} \frac{\mathrm{d}^n}{\mathrm{d} x^n}\left(\mathrm{e}^{-x} x^{n+\alpha}\right)$, $\xi := 2\lambda\mathrm{e}^{-a(r-r_e)}$ are defined. The domain is $0\leq r \lt \infty$.

Antique.ψ — Method

ψ(model::PoschlTeller, x; n::Int=0)

\[\psi_n(x) = \frac{(-1)^μ}{\sqrt{x_0}} P_\lambda^{\mu}(\mathrm{tanh}(x/x_0)) \sqrt{\mu\frac{\Gamma(\lambda-\mu+1)}{\Gamma(\lambda+\mu+1)}},\]

where $\mu = \mu(n) = n_\mathrm{max}-n+1$, and $n_\mathrm{max} = \left\lfloor \lambda \right\rfloor - 1$ and $P_\lambda^{\mu}$ are the associated Legendre functions.

Antique.ψ — Method

ψ(model::RigidRotor, θ, φ; l::Int=0, m::Int=0)

\[\psi_{lm}(\theta,\varphi) = Y_{lm}(\theta,\varphi)\]

The wave functions are the spherical harmonics. The domain is $0\leq \theta \lt \pi$ and $0\leq \varphi \lt 2\pi$.

Antique.ψ — Method

ψ(model::SphericalOscillator, r, θ, φ; n::Int=0, l::Int=0, m::Int=0)

\[\psi_{nlm}(\pmb{r}) = R_{nl}(r) Y_{lm}(\theta,\varphi)\]

The domain is $0\leq r \lt \infty, 0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$.