Isotropic Harmonic Oscillator 3D

The isotropic harmonic oscillator is the most frequently used model in quantum physics.

Definitions

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(r) = E \psi(xr),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2m} \nabla^2 + V(r).\]

Parameters are specified with the following struct.

Parameters

Antique.HarmonicOscillator3D — Type

HarmonicOscillator(k=1.0, m=1.0, ℏ=1.0)

$k$ is the force constant, $m$ is the mass of particle and $\hbar$ is the reduced Planck constant (Dirac's constant).

Potential

Antique.V — Method

V(model::HarmonicOscillator3D, r)

\[V(r) = \frac{1}{2} k r^2 = \frac{1}{2} m \omega^2 r^2 = \frac{1}{2} \hbar \omega \xi^2,\]

where $\omega = \sqrt{k/m}$ is the angular frequency and $\xi = \sqrt{\frac{m\omega}{\hbar}}r$.

Eigen Values

Antique.E — Method

E(model::HarmonicOscillator3D; n=1)

\[E_n = \left(n + \frac{3}{2}\right)\hbar \omega,\]

where $\omega = \sqrt{k/m}$

Eigen Functions

Antique.ψ — Method

ψ(model::HarmonicOscillator3D, r, θ, φ; n=1, l=0, m=0)

\[\psi_{nlm}(\pmb{r}) = R_{nl}(r) Y_{lm}(\theta,\varphi)\]

The domain is $0\leq r \lt \infty, 0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$.

Radial Functions

Antique.R — Method

R(model::HarmonicOscillator3D, r; n=1, l=0)

\[R_{nl}(r) = \sqrt{ \frac{\gamma^{3/2}}{(2\sqrt{\pi}}} \sqrt{\frac{2^{n+l+3} n!}{(2n+2l+1)!!}} \xi^l \exp\left(-\xi^2/2\right)L_{n}^{l+\frac{1}{2}} \left(\xi^2\right),\]

where Laguerre polynomials are defined as $L_n(x) = \frac{1}{n!} \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$, and associated Laguerre polynomials are defined as $L_n^{k}(x) = \frac{\mathrm{d}^k}{\mathrm{d}x^k} L_n(x)$. Note that replace $2n(n+l)!$ with $2n[(n+l)!]^3$ if Laguerre polynomials are defined as $L_n(x) = \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$. The domain is $0\leq r \lt \infty$.

Generalized Laguerre Polynomials

Note

The generalized Laguerre polynomials $L_n^{(\alpha)}(x)$, not the associated Laguerre polynomials $L_n^{k}(x)$, are used in this model.

Antique.L — Method

L(model::HarmonicOscillator3D, x; n=0, k=0)

Rodrigues' formula & closed-form:

\[\begin{aligned} L_n^{k}(x) &= \frac{\mathrm{d}^k}{\mathrm{d}x^k} L_n(x) \\ &= \frac{\mathrm{d}^k}{\mathrm{d}x^k} \frac{1}{n!} \mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right) \\ &= \sum_{m=0}^{n-k} (-1)^{m+k} \frac{n!}{m!(m+k)!(n-m-k)!} x^m \\ &= (-1)^k L_{n-k}^{(k)}(x), \end{aligned}\]

where Laguerre polynomials are defined as $L_n(x)=\frac{1}{n!}\mathrm{e}^x \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left( \mathrm{e}^{-x} x^n \right)$.

Examples:

\[\begin{aligned} L_0^0(x) &= 1, \\ L_1^0(x) &= 1 - x, \\ L_1^1(x) &= 1, \\ L_2^0(x) &= 1 - 2 x + 1/2 x^2, \\ L_2^1(x) &= 2 - x, \\ L_2^2(x) &= 1, \\ L_3^0(x) &= 1 - 3 x + 3/2 x^2 - 1/6 x^3, \\ L_3^1(x) &= 3 - 3 x + 1/2 x^2, \\ L_3^2(x) &= 3 - x, \\ L_3^3(x) &= 1, \\ L_4^0(x) &= 1 - 4 x + 3 x^2 - 2/3 x^3 + 5/12 x^4, \\ L_4^1(x) &= 4 - 6 x + 2 x^2 - 1/6 x^3, \\ L_4^2(x) &= 6 - 4 x + 1/2 x^2, \\ L_4^3(x) &= 4 - x, \\ L_4^4(x) &= 1, \\ \vdots \end{aligned}\]

Spherical Harmonics

Antique.Y — Method

Y(model::HarmonicOscillator3D, θ, φ; l=0, m=0)

\[Y_{lm}(\theta,\varphi) = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{2l+1}{4\pi} \frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} (\cos\theta) \mathrm{e}^{im\varphi}.\]

The domain is $0\leq \theta \lt \pi, 0\leq \varphi \lt 2\pi$. Note that some variants are connected by

\[i^{|m|+m} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^{\frac{|m|+m}{2}} \sqrt{\frac{(l-|m|)!}{(l+|m|)!}} P_l^{|m|} = (-1)^m \sqrt{\frac{(l-m)!}{(l+m)!}} P_l^{m}.\]

Associated Legendre Polynomials

Antique.P — Method

P(model::HarmonicOscillator3D, x; n=0, m=0)

Rodrigues' formula & closed-form:

\[\begin{aligned} P_n^m(x) &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} P_n(x) \\ &= \left( 1-x^2 \right)^{m/2} \frac{\mathrm{d}^m}{\mathrm{d}x^m} \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right] \\ &= \frac{1}{2^n} (1-x^2)^{m/2} \sum_{j=0}^{\left\lfloor\frac{n-m}{2}\right\rfloor} (-1)^j \frac{(2n-2j)!}{j! (n-j)! (n-2j-m)!} x^{(n-2j-m)}. \end{aligned},\]

where Legendre polynomials are defined as $P_n(x) = \frac{1}{2^n n!} \frac{\mathrm{d}^n}{\mathrm{d}x ^n} \left[ \left( x^2-1 \right)^n \right]$. Note that $P_l^{-m} = (-1)^m \frac{(l-m)!}{(l+m)!} P_l^m$ for $m<0$. (It is not compatible with $P_k^m(t) = (-1)^m\left( 1-t^2 \right)^{m/2} \frac{\mathrm{d}^m P_k(t)}{\mathrm{d}t^m}$ caused by $(-1)^m$.) The specific formulae are given below.

Examples:

\[\begin{aligned} P_{0}^{0}(x) &= 1, \\ P_{1}^{0}(x) &= x, \\ P_{1}^{1}(x) &= \left(+1\right)\sqrt{1-x^2}, \\ P_{2}^{0}(x) &= -1/2 + 3/2 x^{2}, \\ P_{2}^{1}(x) &= \left(-3 x\right)\sqrt{1-x^2}, \\ P_{2}^{2}(x) &= 3 - 6 x, \\ P_{3}^{0}(x) &= -3/2 x + 5/2 x^{3}, \\ P_{3}^{1}(x) &= \left(3/2 - 15/2 x^{2}\right)\sqrt{1-x^2}, \\ P_{3}^{2}(x) &= 15 x - 30 x^{2}, \\ P_{3}^{3}(x) &= \left(15 - 30 x\right)\sqrt{1-x^2}, \\ P_{4}^{0}(x) &= 3/8 - 15/4 x^{2} + 35/8 x^{4}, \\ P_{4}^{1}(x) &= \left(- 15/2 x + 35/2 x^{3}\right)\sqrt{1-x^2}, \\ P_{4}^{2}(x) &= -15/2 + 15 x + 105/2 x^{2} - 105 x^{3}, \\ P_{4}^{3}(x) &= \left(105 x - 210 x^{2}\right)\sqrt{1-x^2}, \\ P_{4}^{4}(x) &= 105 - 420 x + 420 x^{2}, \\ & \vdots \end{aligned}\]

Usage & Examples

Install Antique.jl for the first use and run using Antique before each use. The energy E(), wavefunction ψ(), potential V() and some other functions are suppoted. In this system, the model is generated by HarmonicOscillator3D and several parameters k, m and ℏ are set as optional arguments.

using Antique
HO3D = HarmonicOscillator3D(k=1.0, m=1.0, ℏ=1.0)

Parameters:

julia> HO3D.k1.0
julia> HO3D.m1.0
julia> HO3D.ℏ1.0

Eigen values:

julia> E(HO3D, n=0)1.5
julia> E(HO3D, n=1)2.5

Testing

Unit testing and Integration testing were done using computer algebra system (Symbolics.jl) and numerical integration (QuadGK.jl). The test script is here.

Normalization of $R_{nl}(r)$

\[\int |R_{nl}(r)|^2 r^2 \mathrm{d}r = 1\]

 n |  l |        analytical |         numerical 
-- | -- | ----------------- | ----------------- 
 0 |  0 |    1.000000000000 |    1.000000000000 ✔
 1 |  0 |    1.000000000000 |    1.000000000000 ✔
 1 |  1 |    1.000000000000 |    1.000000000000 ✔
 2 |  0 |    1.000000000000 |    1.000000000000 ✔
 2 |  1 |    1.000000000000 |    1.000000000000 ✔
 2 |  2 |    1.000000000000 |    1.000000000000 ✔
 3 |  0 |    1.000000000000 |    1.000000000000 ✔
 3 |  1 |    1.000000000000 |    1.000000000000 ✔
 3 |  2 |    1.000000000000 |    1.000000000000 ✔
 3 |  3 |    1.000000000000 |    1.000000000000 ✔
 4 |  0 |    1.000000000000 |    1.000000000000 ✔
 4 |  1 |    1.000000000000 |    1.000000000000 ✔
 4 |  2 |    1.000000000000 |    1.000000000000 ✔
 4 |  3 |    1.000000000000 |    1.000000000000 ✔
 4 |  4 |    1.000000000000 |    1.000000000000 ✔
 5 |  0 |    1.000000000000 |    1.000000000000 ✔
 5 |  1 |    1.000000000000 |    1.000000000000 ✔
 5 |  2 |    1.000000000000 |    1.000000000000 ✔
 5 |  3 |    1.000000000000 |    1.000000000000 ✔
 5 |  4 |    1.000000000000 |    1.000000000000 ✔
 5 |  5 |    1.000000000000 |    1.000000000000 ✔

Normalization & Orthogonality of $\psi_n(r,\theta,\varphi)$

\[\int \psi_i^\ast(r,\theta,\varphi) \psi_j(r,\theta,\varphi) r^2 \sin(\theta) \mathrm{d}r \mathrm{d}\theta \mathrm{d}\varphi = \delta_{ij}\]

n₁ | n₂ | l₁ | l₂ | m₁ | m₂ |        analytical |         numerical 
-- | -- | -- | -- | -- | -- | ----------------- | ----------------- 
 0 |  0 |  0 |  0 |  0 |  0 |    1.000000000000 |    1.000000000026 ✔
 0 |  1 |  0 |  0 |  0 |  0 |    0.000000000000 |   -0.000000000068 ✔
 0 |  1 |  0 |  1 |  0 | -1 |    0.000000000000 |   -0.000000000000 ✔
 0 |  1 |  0 |  1 |  0 |  0 |    0.000000000000 |   -0.000000000000 ✔
 0 |  1 |  0 |  1 |  0 |  1 |    0.000000000000 |    0.000000000000 ✔
 1 |  0 |  0 |  0 |  0 |  0 |    0.000000000000 |   -0.000000000068 ✔
 1 |  0 |  1 |  0 | -1 |  0 |    0.000000000000 |    0.000000000000 ✔
 1 |  0 |  1 |  0 |  0 |  0 |    0.000000000000 |   -0.000000000000 ✔
 1 |  0 |  1 |  0 |  1 |  0 |    0.000000000000 |    0.000000000000 ✔
 1 |  1 |  0 |  0 |  0 |  0 |    1.000000000000 |    0.999999995268 ✔
 1 |  1 |  0 |  1 |  0 | -1 |    0.000000000000 |   -0.000000000000 ✔
 1 |  1 |  0 |  1 |  0 |  0 |    0.000000000000 |   -0.000000000000 ✔
 1 |  1 |  0 |  1 |  0 |  1 |    0.000000000000 |    0.000000000000 ✔
 1 |  1 |  1 |  0 | -1 |  0 |    0.000000000000 |   -0.000000000000 ✔
 1 |  1 |  1 |  0 |  0 |  0 |    0.000000000000 |   -0.000000000000 ✔
 1 |  1 |  1 |  0 |  1 |  0 |    0.000000000000 |    0.000000000000 ✔
 1 |  1 |  1 |  1 | -1 | -1 |    1.000000000000 |    0.999999951768 ✔
 1 |  1 |  1 |  1 | -1 |  0 |    0.000000000000 |   -0.000000000000 ✔
 1 |  1 |  1 |  1 | -1 |  1 |    0.000000000000 |   -0.000000000000 ✔
 1 |  1 |  1 |  1 |  0 | -1 |    0.000000000000 |   -0.000000000000 ✔
 1 |  1 |  1 |  1 |  0 |  0 |    1.000000000000 |    0.999999951768 ✔
 1 |  1 |  1 |  1 |  0 |  1 |    0.000000000000 |    0.000000000000 ✔
 1 |  1 |  1 |  1 |  1 | -1 |    0.000000000000 |   -0.000000000000 ✔
 1 |  1 |  1 |  1 |  1 |  0 |    0.000000000000 |    0.000000000000 ✔
 1 |  1 |  1 |  1 |  1 |  1 |    1.000000000000 |    0.999999951768 ✔