Infinite Potential Well 3D (Particle in a 3D-Box)

The three-dimensional infinite potential well (particle in a 3D-box) is one of the simplest models for a three-dimensional quantum mechanical system.

Definitions

This model is described with the time-independent Schrödinger equation

\[ \hat{H} \psi(x,y,z) = E \psi(x,y,z),\]

and the Hamiltonian

\[ \hat{H} = - \frac{\hbar^2}{2m} \left(\frac{\partial^2}{\partial x ^2} + \frac{\partial^2}{\partial y ^2} + \frac{\partial^2}{\partial z ^2}\right) + V(x,y,z).\]

Parameters are specified with the following struct.

Parameters

Antique.InfinitePotentialWell3D — Type

InfinitePotentialWell3D(Lx=1.0, Ly=1.0, Lz=1.0, m=1.0, ℏ=1.0)

$L_x,L_y,L_z$ are the lengths of the box in $x$,$y$,$z$-direction, $m$ is the mass of particle and $\hbar$ is the reduced Planck constant (Dirac's constant).

Potential

Antique.V — Method

V(model::InfinitePotentialWell3D, x,y,z)

\[V(x,y,z) = \left\{ \begin{array}{ll} 0 & 0 \leq x \leq L_x \ \mathrm{and}\ 0 \leq y \leq L_y \ \mathrm{and}\ 0 \leq z \leq L_z \\ \infty & \mathrm{elsewhere} \end{array} \right.\]

Eigen Values

Antique.E — Method

E(model::InfinitePotentialWell3D; nx=1, ny=1, nz=1)

\[E_{n_x,n_y,n_z} = \frac{\hbar^2 \pi^2}{2 m} \left(\frac{n_x^2}{L_x^2} + \frac{n_y^2}{L_y^2} + \frac{n_z^2}{L_z^2}\right)\]

Eigen Functions

Antique.ψ — Method

ψ(model::InfinitePotentialWell3D, x,y,z; nx=1, ny=1, nz=1)

The wave functions can be expressed as products of wave functions in a one-dimensional box.

\[\psi_{n_x,n_y,n_z}(x,y,z) = \psi_{n_x}(x)\psi_{n_y}(y)\psi_{n_z}(z) = \sqrt{\frac{8}{L_xL_yL_z}} \sin\left(\frac{n_x\pi x}{L_x}\right) \sin\left(\frac{n_y\pi y}{L_y}\right) \sin\left(\frac{n_z\pi z}{L_z}\right)\]

Usage & Examples

Install Antique.jl for the first use and run using Antique before each use. The energy E(), wavefunction ψ(), potential V() and some other functions are supported. In this system, the model is generated by InfinitePotentialWell3D and several parameters Lx, Ly, Lz, m and ℏ are set as optional arguments.

using Antique
IPW3D = InfinitePotentialWell3D(Lx=1.0, Ly=1.0, Lz=1.0, m=1.0, ℏ=1.0)

Parameters:

julia> IPW3D.Lx1.0
julia> IPW3D.Ly1.0
julia> IPW3D.Lz1.0
julia> IPW3D.m1.0
julia> IPW3D.ℏ1.0

Eigen values:

julia> E(IPW3D, nx=1, ny=1, nz=1)14.804406601634037
julia> E(IPW3D, nx=2, ny=1, nz=1)29.608813203268074
julia> E(IPW3D, nx=1, ny=2, nz=1)29.608813203268074
julia> E(IPW3D, nx=1, ny=1, nz=2)29.608813203268074
julia> E(IPW3D, nx=2, ny=2, nz=1)44.41321980490211
julia> E(IPW3D, nx=2, ny=1, nz=2)44.41321980490211
julia> E(IPW3D, nx=1, ny=2, nz=2)44.41321980490211
julia> E(IPW3D, nx=2, ny=2, nz=2)59.21762640653615

Wave functions:

The wave functions of the 3D infinite potential well can be expressed as a product of wave functions in 1D infinite potential wells for each of the three directions $x$,$y$,$z$. As the representation of 3D (non-spherical) wave functions easily becomes unclear, we refer to the documentation of the one-dimensional counterpart: InfinitePotentialWell (...)

Testing

Unit testing and Integration testing were done using numerical integration (QuadGK.jl). The test script is here.

Normalization & Orthogonality of $\psi_{n_x,n_y,n_z}(x,y,z)$

\[\int_{0}^{L_x}\int_{0}^{L_y}\int_{0}^{L_z} \psi_{i_x,i_y,i_z}^\ast(x,y,z) \psi_{j_x,j_y,j_z}(x,y,z) ~\mathrm{d}x \mathrm{d}y\mathrm{d}z = \delta_{i_x,j_x}\delta_{i_y,j_y}\delta_{i_z,j_z}\]

ix | iy | iz | jx | jy | jz |        analytical |         numerical 
-- | -- | -- | -- | -- | -- | ----------------- | ----------------- 

Eigen Values

\[ \begin{aligned} E_n &= \int_0^L \psi^\ast_n(x) \hat{H} \psi_n(x) ~\mathrm{d}x \\ &= \int_0^L \psi^\ast_n(x) \left[ \hat{V} + \hat{T} \right] \psi(x) ~\mathrm{d}x \\ &= \int_0^L \psi^\ast_n(x) \left[ 0 - \frac{\hbar^2}{2m} \frac{\mathrm{d}^{2}}{\mathrm{d} x^{2}} \right] \psi(x) ~\mathrm{d}x \\ &\simeq \int_0^L \psi^\ast_n(x) \left[ -\frac{\hbar^2}{2m} \frac{\psi(x+\Delta x) - 2\psi(x) + \psi(x-\Delta x)}{\Delta x^{2}} \right] ~\mathrm{d}x. \end{aligned}\]

Where, the difference formula for the 2nd-order derivative:

\[\begin{aligned} % 2\psi(x) % + \frac{\mathrm{d}^{2} \psi(x)}{\mathrm{d} x^{2}} \Delta x^{2} % + O\left(\Delta x^{4}\right) % &= % \psi(x+\Delta x) % + \psi(x-\Delta x) % \\ % \frac{\mathrm{d}^{2} \psi(x)}{\mathrm{d} x^{2}} \Delta x^{2} % &= % \psi(x+\Delta x) % - 2\psi(x) % + \psi(x-\Delta x) % - O\left(\Delta x^{4}\right) % \\ % \frac{\mathrm{d}^{2} \psi(x)}{\mathrm{d} x^{2}} % &= % \frac{\psi(x+\Delta x) - 2\psi(x) + \psi(x-\Delta x)}{\Delta x^{2}} % - \frac{O\left(\Delta x^{4}\right)}{\Delta x^{2}} % \\ \frac{\mathrm{d}^{2} \psi(x)}{\mathrm{d} x^{2}} &= \frac{\psi(x+\Delta x) - 2\psi(x) + \psi(x-\Delta x)}{\Delta x^{2}} + O\left(\Delta x^{2}\right) \end{aligned}\]

are given by the sum of 2 Taylor series:

\[\begin{aligned} \psi(x+\Delta x) &= \psi(x) + \frac{\mathrm{d} \psi(x)}{\mathrm{d} x} \Delta x + \frac{1}{2!} \frac{\mathrm{d}^{2} \psi(x)}{\mathrm{d} x^{2}} \Delta x^{2} + \frac{1}{3!} \frac{\mathrm{d}^{3} \psi(x)}{\mathrm{d} x^{3}} \Delta x^{3} + O\left(\Delta x^{4}\right), \\ \psi(x-\Delta x) &= \psi(x) - \frac{\mathrm{d} \psi(x)}{\mathrm{d} x} \Delta x + \frac{1}{2!} \frac{\mathrm{d}^{2} \psi(x)}{\mathrm{d} x^{2}} \Delta x^{2} - \frac{1}{3!} \frac{\mathrm{d}^{3} \psi(x)}{\mathrm{d} x^{3}} \Delta x^{3} + O\left(\Delta x^{4}\right). \end{aligned}\]

 nx |  ny |  nz |        analytical |         numerical 
 -- | --- | --- | ----------------- | -----------------